updated wishlist logic

+++ src/main/java/com/example/jwtauthorisedlogin/service/WishlistService.java
@@ -2,17 +2,22 @@ package com.example.jwtauthorisedlogin.service;
 
 import com.example.jwtauthorisedlogin.Entity.Food;
 import com.example.jwtauthorisedlogin.Entity.UserWishlist;
-import com.example.jwtauthorisedlogin.payload.request.WishListGetRequest;
+//import com.example.jwtauthorisedlogin.payload.request.WishListGetRequest;
+import com.example.jwtauthorisedlogin.payload.request.WishlistDeleteRequest;
 import com.example.jwtauthorisedlogin.payload.request.WishlistRequest;
 import com.example.jwtauthorisedlogin.payload.response.MessageResponse;
 import com.example.jwtauthorisedlogin.repository.FoodRepository;
 import com.example.jwtauthorisedlogin.repository.UserRepository;
 import com.example.jwtauthorisedlogin.repository.UserWishlistRepository;
+import com.example.jwtauthorisedlogin.user.User;
 import lombok.AllArgsConstructor;
 import org.springframework.data.crossstore.ChangeSetPersister;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -24,16 +29,18 @@ public class WishlistService {
     private final FoodRepository foodRepository;
 
     public MessageResponse addToWishlist(WishlistRequest request) {
-            var user = userRepository.findByEmail(request.getUserEmail()).orElse(null);
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            User currentUser = (User) authentication.getPrincipal();
+
             var food = foodRepository.findById(request.getFoodId()).orElse(null);
 
-            if (wishlistRepository.existsByUserAndFood(user, food)) {
+            if (wishlistRepository.existsByUserAndFood(currentUser, food)) {
                 return MessageResponse.builder().message("Food item is already in the wishlist").build();
             }
 
 
             var wishlistItem = UserWishlist.builder()
-                .user(user)
+                .user(currentUser)
                 .food(food)
                 .build();
 
@@ -42,26 +49,40 @@ public class WishlistService {
         return MessageResponse.builder().message("Food item added to the wishlist").build();
     }
 
-    public List<Food> getWishlistItems(WishListGetRequest request) {
-        return wishlistRepository.findByUserEmail(request.getUserEmail())
+    public List<Food> getWishlistItems() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User currentUser = (User) authentication.getPrincipal();
+        var user = userRepository.findByEmail(currentUser.getEmail()).orElse(null);
+
+        assert user != null;
+        return wishlistRepository.findByUserEmail(user.getEmail())
                 .stream()
                 .map(UserWishlist::getFood)
                 .collect(Collectors.toList());
     }
 
-    public MessageResponse deleteWishlistItem(WishlistRequest request) {
-        var user = userRepository.findByEmail(request.getUserEmail()).orElse(null);
+    public MessageResponse deleteWishlistItem(WishlistDeleteRequest request) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User currentUser = (User) authentication.getPrincipal();
+        var user = userRepository.findByEmail(request.getEmail()).orElse(null);
         var food = foodRepository.findById(request.getFoodId()).orElse(null);
 
-        if (!wishlistRepository.existsByUserAndFood(user, food)) {
+        if(Objects.equals(currentUser.getEmail(), request.getEmail())){
+            if (!wishlistRepository.existsByUserAndFood(user,food)) {
+                return MessageResponse.builder().message("Food item is not in the wishlist").build();
+            }
+
+            var wishlistItem = wishlistRepository.findByUserAndFood(user, food).orElse(null);
+
+            assert wishlistItem != null;
+            wishlistRepository.delete(wishlistItem);
+
+            return MessageResponse.builder().message("Food item deleted from the wishlist").build();
+        }
+        else{
             return MessageResponse.builder().message("Food item is not in the wishlist").build();
         }
 
-        var wishlistItem = wishlistRepository.findByUserAndFood(user, food).orElse(null);
-
-        wishlistRepository.delete(wishlistItem);
-
-        return MessageResponse.builder().message("Food item deleted from the wishlist").build();
     }
 
 }
